refactor(DropdownContext): extract to dedicated file
