+++ libs/spark/src/DropdownContext.tsx
@@ -1,0 +1,52 @@
+import * as React from 'react';
+import { useUniqueId } from './utils';
+
+// This file is an adaption of Mui's TabContext.
+// See https://github.com/mui-org/material-ui/blob/cd8aeb64a9038ae9cb8ecf1cd7fd8a2bfea2d6b6/packages/material-ui-lab/src/TabContext/TabContext.js
+
+export interface DropdownContextValue {
+  id: string;
+  anchorEl: null | HTMLElement;
+  handleClick: (event: React.MouseEvent<HTMLElement>) => void;
+  handleClose: (event: React.MouseEvent<HTMLElement>) => void;
+}
+
+const Context = React.createContext<DropdownContextValue | null>(null);
+
+if (process.env.NODE_ENV !== 'production') {
+  Context.displayName = 'DropdownContext';
+}
+
+export interface DropdownContextProps {
+  /**
+   * The content of the component.
+   */
+  children?: React.ReactNode;
+}
+
+export default function DropdownContext(
+  props: DropdownContextProps
+): JSX.Element {
+  const id = useUniqueId();
+
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+
+  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  const value = React.useMemo(
+    () => ({ id, anchorEl, handleClick, handleClose }),
+    [id, anchorEl]
+  );
+
+  return <Context.Provider value={value} {...props} />;
+}
+
+export function useDropdownContext(): DropdownContextValue | null {
+  return React.useContext(Context);
+}
